requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
boto3==1.34.108; python_version >= '3.8'
botocore==1.34.108; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
channels==4.1.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
cryptography==42.0.7; python_version >= '3.7'
django==5.0.6; python_version >= '3.10'
django-redis==5.4.0; python_version >= '3.6'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-jwt==1.11.0
frozenlist==1.4.1; python_version >= '3.8'
gevent==24.2.1; python_version >= '3.8'
greenlet==3.0.3; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
hexbytes==1.2.0; python_version >= '3.8' and python_version < '4'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
multidict==6.0.5; python_version >= '3.7'
pika==1.3.2; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycparser==2.22; python_version >= '3.8'
pydantic==2.7.1; python_version >= '3.8'
pydantic-core==2.18.2; python_version >= '3.8'
pyjwt==1.7.1
pymysql==1.1.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
redis==5.0.4; python_version >= '3.7'
s3transfer==0.10.1; python_version >= '3.8'
setuptools==69.5.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.11.0; python_version >= '3.8'
tzdata==2024.1; sys_platform == 'win32'
urllib3==2.2.1; python_version >= '3.10'
uvicorn==0.29.0; python_version >= '3.8'
web3==6.19.0
websocket==0.2.1
yarl==1.9.4; python_version >= '3.7'
zope.event==5.0; python_version >= '3.7'
zope.interface==6.4; python_version >= '3.7'